requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-phonenumber-field==5.0.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonfield==3.1.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
MarkupSafe==1.1.1
mistune==0.8.4
mpmath==1.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
oauthlib==3.1.0
packaging==20.7
pandocfilters==1.4.3
parso==0.7.1
phonenumbers==8.12.9
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.2
PyJWT==2.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
reportlab==3.5.66
requests==2.25.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
sympy==1.6.2
terminado==0.9.1
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.26.2
virtualenv==20.2.1
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
